reseed
added a bestglm for an optimized model based on a new randomized sample that is twice the size of any prior sample.  So going in one would have an idea of the # of resamples that would be required to infer a population statistic without having to derive it from the entire population.  In this case, we choose a 2% training/holdout (4%) which is resampled twice (8%) .  At the end, we add another 2% training/holdout (4%), which results in 12% total