tempted to move block outside of code
no prior code changes which is good.  Considering moving the post 3 aggregate to outside of file iterator.  As is, I need to have the data prepped into memory or loaded from csv.  Seriously considering a folder that is used for output/input to wipe/reload all csv's and process (rather than parsing names)