Update cleandatacode.R
added resampling.  Realized I WANT to use the sample data.train for each category as I can insert a newList just before I called resampleMC which will utilize it.  Which means I'm not resusing my sample other than for different categories and am still reserving my holdout partition for filtering at the end of it all.  Yes, there are excessive loops where data.train is used, but that's okay, data.train is used differently per category.  Then at the end data.test is used to filter data.train which will have not have any information that was in data.train.  This is important because essentially while the output is different >= median vs an actual best cv subset.  Is that they both build the same model of the same data if I used merged... so I need to not use merged...  I need to use data.test to have a separate model to do the pruning by.  Unfortunately this is more expensive (would be cheaper to resample and merge data.train/data.test but then I would be resuing information).