reverted class balancing to size of class 1
this is because class 1 is generally the class of interest and is rare.  We want to know what makes it presence come into existence (hence, modify y).  So we sample to fit IT's set of observations.

I did setup more balanced classes at the lower strata (MC Resample).  But considering what I just did balances it at the top, it's kind of redundant, but it does make for faster computations and ensures a 50/50 split at the lowest level