major rework on proportional class redbalancing
under/oversamples based on mean(%[ones,zeros]) that is divided by the min(%[ones,zeros]) to derive a reloop factor as well as remainder.  This balances the two classes around the average. I believe the error I've been getting with bestglm is due to too large a dataframe to work with.  I saw similar errors in newCSVs.R so I've been working on a monte carlo resample technique along with sub_CVnames to filter out a set vs the best subset.